product issue updated

+++ app/Controller/ProductController.php
@@ -9,7 +9,12 @@ class ProductController extends Controller {
 
     public function index() {
         $results = $this->selectAll('products')->fetchAll(\PDO::FETCH_ASSOC);
-        return json_encode(['status'=> true, 'data'=>$results]);
+        $resultset = [];
+        foreach ($results as $result) {
+            $result['image_url'] = 'http://'.$_SERVER['HTTP_HOST'].'/storage/uploads/'.$result['image'];
+            $resultset[] = $result;
+        }
+        return json_encode(['status'=> true, 'data'=>$resultset]);
     }
 
 
@@ -33,6 +38,18 @@ class ProductController extends Controller {
                 'data'   => 'sku is required' 
             ]);
         }
+        if ($_FILES['image']['error'] == 0) {
+            $dir = $_SERVER['DOCUMENT_ROOT'].'/storage/uploads/';
+            $returnedData = imageupload('image', $dir);
+            if ($returnedData['status']) {
+                $dataLog['image'] = $returnedData['data'];
+            } else {
+                $dataLog['image'] = null;
+            }
+        } else {
+            $dataLog['image'] = null;
+        }
+
         $dataLog['created_at'] = date('Y-m-d H:i:s');
         $stored = $this->insert('products', $dataLog);
         $dataLog['id'] = $this->getInsertId();
@@ -66,35 +83,65 @@ class ProductController extends Controller {
     }
 
     public function updateData($dataLog) {
-        if (isset($dataLog['id']) && empty($dataLog['id'])) {
-            return json_encode([
-                'status' => false,
-                'data'   => 'ID is required' 
-            ]);
-        }
-        $dataLog['updated_at'] = date('Y-m-d H:i:s');
-        $updated = $this->update('products', $dataLog, 'id="'.$dataLog['id'].'"');
-        if ($updated) {
-            return json_encode([
-                'status' => true,
-                'data'   => 'Data has been Updated' 
-            ]);
+        try {
+            if (isset($dataLog['id']) && empty($dataLog['id'])) {
+                throw new \Exception("ID is required", 1);
+            }
+            $updatableData = $this->joinQuery('SELECT * FROM products WHERE id="'.$dataLog['id'].'"')->fetch(\PDO::FETCH_ASSOC);
+            if (!isset($updatableData['name'])) {
+                throw new \Exception("Record not exist", 1);
+            }
+            if ($_FILES['image']['error'] == 0) {
+                $dir = $_SERVER['DOCUMENT_ROOT'].'/storage/uploads/';
+                $returnedData = imageupload('image', $dir);
+                if ($returnedData['status']) {
+                    $oldFile = $dir.$updatableData['image'];
+                    if (file_exists($oldFile)) {
+                        unlink($oldFile);
+                    } 
+                    $dataLog['image'] = $returnedData['data'];
+                } else {
+                    $dataLog['image'] = null;
+                }
+            } else {
+                $dataLog['image'] = null;
+            }
+            $dataLog['updated_at'] = date('Y-m-d H:i:s');
+            $updated = $this->update('products', $dataLog, 'id="'.$dataLog['id'].'"');
+            if ($updated) {
+                return json_encode([
+                    'status' => true,
+                    'data'   => 'Data has been Updated' 
+                ]);
+            }
+        } catch (\Exception $e) {
+            return json_encode(['status' => false, 'data'   => $e->getMessage()]);
         }
     }
 
     public function destroy($dataLog) {
-        if (isset($dataLog['id']) && empty($dataLog['id'])) {
-            return json_encode([
-                'status' => false,
-                'data'   => 'ID is required' 
-            ]);
-        }
-        $destroyed = $this->delete('products', 'id="'.$dataLog['id'].'"');
-        if ($destroyed) {
-            return json_encode([
-                'status' => true,
-                'data'   => 'Data has been deleted'
-            ]);
+        try {
+            if (isset($dataLog['id']) && empty($dataLog['id'])) {
+                    throw new \Exception("ID is required", 1);
+            }
+            $updatableData = $this->joinQuery('SELECT * FROM products WHERE id="'.$dataLog['id'].'"')->fetch(\PDO::FETCH_ASSOC);
+            if (!isset($updatableData['name'])) {
+                throw new \Exception("Record not exist", 1);
+            }
+            $dir = $_SERVER['DOCUMENT_ROOT'].'/storage/uploads/';
+            $oldFile = $dir.$updatableData['image'];
+            if (file_exists($oldFile)) {
+                unlink($oldFile);
+            } 
+            $destroyed = $this->delete('products', 'id="'.$dataLog['id'].'"');
+            if ($destroyed) {
+                return json_encode([
+                    'status' => true,
+                    'data'   => 'Data has been deleted'
+                ]);
+            }
+        } catch (\Exception $e) {
+            return json_encode(['status' => false, 'data'   => $e->getMessage()]);
         }
     }
 }
